Refactor: 사진 접근 권한 여부를 확인하는 메서드를 	  DataSource 객체가 하도록 코드 수정함
